arch/arm/src/stm32/stm32_otgfsdev.c/stm32_otghsdev, stm32f7/stm32_otgdev.c, and stm32l4/stm32l4_otgfsdev.c:  Fix a warnings due to a function prototype mismatch.

+++ arch/arm/src/stm32/stm32_otghsdev.c
@@ -605,7 +605,7 @@ static void        stm32_ep_freereq(FAR struct usbdev_ep_s *ep,
 /* Endpoint buffer management */
 
 #ifdef CONFIG_USBDEV_DMA
-static void       *stm32_ep_allocbuffer(FAR struct usbdev_ep_s *ep, unsigned bytes);
+static void       *stm32_ep_allocbuffer(FAR struct usbdev_ep_s *ep, uint16_t bytes);
 static void        stm32_ep_freebuffer(FAR struct usbdev_ep_s *ep, FAR void *buf);
 #endif
 
@@ -4284,7 +4284,7 @@ static void stm32_ep_freereq(FAR struct usbdev_ep_s *ep, FAR struct usbdev_req_s
  ****************************************************************************/
 
 #ifdef CONFIG_USBDEV_DMA
-static void *stm32_ep_allocbuffer(FAR struct usbdev_ep_s *ep, unsigned bytes)
+static void *stm32_ep_allocbuffer(FAR struct usbdev_ep_s *ep, uint16_t bytes)
 {
   usbtrace(TRACE_EPALLOCBUFFER, privep->epphy);
