+++ arch/arm/src/stm32l4/stm32l4_otgfsdev.c
@@ -712,7 +712,7 @@ static void        stm32l4_ep_freereq(FAR struct usbdev_ep_s *ep,
 
 #ifdef CONFIG_USBDEV_DMA
 static void       *stm32l4_ep_allocbuffer(FAR struct usbdev_ep_s *ep,
-                     unsigned bytes);
+                     uint16_t bytes);
 static void        stm32l4_ep_freebuffer(FAR struct usbdev_ep_s *ep,
                      FAR void *buf);
 #endif
@@ -4406,7 +4406,7 @@ static void stm32l4_ep_freereq(FAR struct usbdev_ep_s *ep, FAR struct usbdev_req
  ****************************************************************************/
 
 #ifdef CONFIG_USBDEV_DMA
-static void *stm32l4_ep_allocbuffer(FAR struct usbdev_ep_s *ep, unsigned bytes)
+static void *stm32l4_ep_allocbuffer(FAR struct usbdev_ep_s *ep, uint16_t bytes)
 {
   usbtrace(TRACE_EPALLOCBUFFER, ((FAR struct stm32l4_ep_s *)ep)->epphy);
